fastapi_backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pyjwt
httpx
python-dotenv
passlib[bcrypt]
